requirements.txt
openai
python-dotenv
azure-search-documents
azure-identity
tiktoken
PyPDF2
tqdm